feat(dataplex): move all tests for Aspect Types to single class to reduce number of RPC calls
